jr/requirements.txt
jira>=3.5.0
python-dotenv>=1.0.0
argparse>=1.4.0 